+++ src/mctslib/cpp/move/mcts_iter_move.h
@@ -31,7 +31,7 @@ inline auto mcts_iter_move_rng = []<class Alg>(Alg & alg, typename Alg::Settings
     }
 
     std::vector<std::shared_ptr<typename Alg::Node>> ties {};
-    int max = alg.body.current_node->children[0]->stats.avg_value();
+    double max = alg.body.current_node->children[0]->stats.avg_value();
 
     for (std::shared_ptr<typename Alg::Node> node_ptr : alg.body.current_node->children) {
         if (node_ptr->stats.avg_value() > max) {
